+++ capycli/bom/check_bom.py
@@ -158,7 +158,7 @@ class CheckBom(capycli.common.script_base.ScriptBase):
             logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Check that all releases in the SBOM exist on target SW360 instance.\n")
 
         if args.help: