+++ capycli/bom/map_bom.py
@@ -832,7 +832,7 @@ class MapBom(capycli.common.script_base.ScriptBase):
             logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Map a given SBOM to data on SW360\n")
 
         if args.help: