+++ capycli/bom/merge_bom.py
@@ -122,7 +122,7 @@ class MergeBom(capycli.common.script_base.ScriptBase):
             LOG = capycli.get_logger(__name__)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Merge two SBOM files.\n")
 
         if args.help: