+++ capycli/dependencies/nuget.py
@@ -662,7 +662,7 @@ class GetNuGetDependencies(capycli.common.script_base.ScriptBase):
             logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Determine Nuget components/dependencies\n")
 
         if args.help: