+++ capycli/moverview/moverview_to_html.py
@@ -81,7 +81,7 @@ class MappingOverviewToHtml(capycli.common.script_base.ScriptBase):
             LOG = capycli.get_logger(__name__)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Create a HTML page showing the mapping overview\n")
 
         if args.help: