+++ capycli/project/check_prerequisites.py
@@ -270,7 +270,7 @@ class CheckPrerequisites(capycli.common.script_base.ScriptBase):
             logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Checks whether all prerequisites for a successful software " +
             "clearing are fulfilled\n")