+++ capycli/project/create_bom.py
@@ -182,7 +182,7 @@ class CreateBom(capycli.common.script_base.ScriptBase):
             logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Create a SBOM for a project on SW360\n")
 
         if args.help: