+++ capycli/project/show_ecc.py
@@ -164,7 +164,7 @@ class ShowExportControlStatus(capycli.common.script_base.ScriptBase):
             logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Show project export control details\n")
 
         if args.help: