+++ capycli/project/show_licenses.py
@@ -198,7 +198,7 @@ class ShowLicenses(capycli.common.script_base.ScriptBase):
             logging.getLogger("urllib3.connectionpool").setLevel(logging.WARNING)
 
         print_text(
-            "\n" + capycli.APP_NAME + ", " + capycli.get_app_version() +
+            "\n" + capycli.get_app_signature() +
             " - Show licenses of all cleared components.")
 
         if args.help: